Finish Form
I changed my form from autocomplete off to on for more accessibility and updated my pattern because the regex was set up for first and last name and I wanted to modify it for this forms specific needs. updated dropdown list options for destination and travel time lists so you cant select choose as a real option and it shows up as the default option first.